Fixed issue 2018878. If a SOAP error is received from the server, it will be included in the exception that is thrown back at the end-user.
There is still an issue with some IOException that can reach the end-user layer. This is a no-no and will need to be fixed. I didn't fix it here since the MondrianInprocProxy implements the faulty interface. XmlaOlap4jProxy implementation should not throw any IOException, but should handle it's own exceptions and wrap them in XmlaOlap4jProxyException is necessary.

I've added a FIXME tag and will log a bug against it.

git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@119 c6a108a4-781c-0410-a6c6-c2d559e19af0