I have double checked the dependencies tree and made sure there was nothing circular. I have also uploaded a schema to help you understand better the current dependencies. It's available here :
https://sourceforge.net/tracker2/download.php?group_id=168953&atid=848534&file_id=299997&aid=1991341

I have updated the build file to reflect these changes and added some more
tweaks to it. It is now possible to generate the javadoc of the XML/A
driver, for example.

Before the 1.0 release, we'll need to further "disconnect" all the
components. I'll create a separate task for this.

git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@129 c6a108a4-781c-0410-a6c6-c2d559e19af0