First off I updated the checkFile utility with the most recent version. It now checks for line lenghts and function wrapping.
There were Windows line endings in the Query package. Since all other files were Unix line endings, I converted the query package and now the official olap4j's line ending scheme is Unix. It's not like we had any other choice anyways, checkFile has trouble validating files with Windows line endings on a Linux machine... go figure.)

Also added an improvement to the query package. It now supports exclusions. "Selections" is now a concept that represents a meta group of members rather than members to include in a query. What was previously known as selections should now be referenced to as inclusions.

Simply put, we perform selections and decide if this selection should be included or excluded from the query. This allows the query model, for example, to select all children of a node, and exclude specific ones. The exclusion mechanism uses the MDX Except() function, and exclusions have precedence over inclusions.

I also added a test for that to the OlapTest class.

git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@261 c6a108a4-781c-0410-a6c6-c2d559e19af0