XMLA driver now supports localized connections.
Implementation is as follows. The driver passes connection's locale in as the
LocaleIdentifier property in metadata requests. As of change 14504,
MONDRIAN-367, mondrian observes the LocaleIdentifier of XMLA requests, and the
Sales cube in the FoodMart schema is localized into French and German. Since
each metadata element only contains caption and description of current locale,
switching locale causes metadata cache flush.)

Revert unintended change to buildJdk16.sh.


git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@470 c6a108a4-781c-0410-a6c6-c2d559e19af0